add pagination mapper and change aired to string #JC19-352
